+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/main/java/com/google/ads/mediation/vungle/VungleMediationAdapter.java
@@ -34,6 +34,7 @@ import com.google.ads.mediation.vungle.rtb.VungleRtbNativeAd;
 import com.google.ads.mediation.vungle.rtb.VungleRtbRewardedAd;
 import com.google.ads.mediation.vungle.waterfall.VungleWaterfallAppOpenAd;
 import com.google.ads.mediation.vungle.waterfall.VungleWaterfallBannerAd;
+import com.google.ads.mediation.vungle.waterfall.VungleWaterfallInterstitialAd;
 import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.VersionInfo;
@@ -77,6 +78,7 @@ public class VungleMediationAdapter extends RtbAdapter
   public static final String TAG = VungleMediationAdapter.class.getSimpleName();
 
   private VungleWaterfallAppOpenAd waterfallAppOpenAd;
+  private VungleWaterfallInterstitialAd waterfallInterstitialAd;
 
   private VungleRtbBannerAd rtbBannerAd;
   private VungleRtbInterstitialAd rtbInterstitialAd;
@@ -281,6 +283,18 @@ public class VungleMediationAdapter extends RtbAdapter
     VungleWaterfallBannerAd waterfallBannerAd =
         new VungleWaterfallBannerAd(callback, vungleFactory);
     waterfallBannerAd.validateParamsAndLoadAd(mediationBannerAdConfiguration);
+  }
+
+  @Override
+  public void loadInterstitialAd(
+      @NonNull MediationInterstitialAdConfiguration mediationInterstitialAdConfiguration,
+      @NonNull
+          MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback>
+              callback) {
+    VungleInitializer.getInstance()
+        .updateCoppaAndUnderageConsentStatus(MobileAds.getRequestConfiguration());
+    waterfallInterstitialAd = new VungleWaterfallInterstitialAd(callback, vungleFactory);
+    waterfallInterstitialAd.render(mediationInterstitialAdConfiguration);
   }
 
   @Override